move common config.go to a separate folder common/config/config.go
