+++ pkg/common/config/config.go
@@ -1,4 +1,4 @@
-package common
+package config
 
 import (
 	"fmt"
